if file doesnt exist copy workload to it

+++ ci-operator/step-registry/openshift-qe/cluster-density-v2/openshift-qe-cluster-density-v2-commands.sh
@@ -38,5 +38,11 @@ folder_name=$(ls -t -d /tmp/*/ | head -1)
 jq ".iterations = $ITERATIONS" $folder_name/index_data.json >> ${SHARED_DIR}/index_data.json
 
 echo "{\"$WORKLOAD\": \"$ITERATIONS\"}" > workload.json 
-result=$(jq -s add workload.json ${SHARED_DIR}/perfscale_run.json)
-echo $result > ${SHARED_DIR}/perfscale_run.json
+
+if [ -f "${SHARED_DIR}/perfscale_run.json" ]; then
+    result=$(jq -s add workload.json ${SHARED_DIR}/perfscale_run.json)
+    echo $result > ${SHARED_DIR}/perfscale_run.json
+else
+    cp workload.json ${SHARED_DIR}/perfscale_run.json
+fi
+
