+++ ci-operator/step-registry/openshift-qe/network-policy/openshift-qe-network-policy-commands.sh
@@ -51,5 +51,9 @@ folder_name=$(ls -t -d /tmp/*/ | head -1)
 jq ".iterations = $JOB_ITERATIONS" $folder_name/index_data.json >> ${SHARED_DIR}/index_data.json
 
 echo "{\"$WORKLOAD\": \"$JOB_ITERATIONS\"}" > workload.json 
-result=$(jq -s add workload.json ${SHARED_DIR}/perfscale_run.json)
-echo $result > ${SHARED_DIR}/perfscale_run.json
+if [ -f "${SHARED_DIR}/perfscale_run.json" ]; then
+    result=$(jq -s add workload.json ${SHARED_DIR}/perfscale_run.json)
+    echo $result > ${SHARED_DIR}/perfscale_run.json
+else
+    cp workload.json ${SHARED_DIR}/perfscale_run.json
+fi
